[GH-89] Clean unused
