+++ src/apis/subject.ts
@@ -1,5 +1,9 @@
 import { http } from 'libs/http'
 
+// =============
+// CRUD Endpoint
+// =============
+
 export interface ISubject {
   id: string
   code: string
@@ -12,44 +16,6 @@ export interface ISubject {
   curriculumCode: string
   isInter: boolean
 }
-
-export interface ICompensated {
-  section: number
-  compensatedList: {
-    compensatedId: string
-    originalDate: string
-    originalTimeList: { start: string; end: string }[]
-    compensatedDate: string
-    compensatedTimeList: { start: string; end: string }[]
-    originalRoom: string
-    compensatedRoom: string
-  }[]
-}
-
-// ==================
-// Subject x Workload
-// ==================
-
-export const getManyCompensatedOfSubject = async (
-  subjectId: string,
-  query: any
-) => {
-  const { data } = await http.get<ICompensated[]>(
-    `/subject/${subjectId}/compensation-workload`,
-    {
-      params: query,
-    }
-  )
-  return data
-}
-
-export const deleteOneCompensated = async (compensatedId: string) => {
-  await http.delete(`/subject/compensated/${compensatedId}`)
-}
-
-// =============
-// CRUD Endpoint
-// =============
 
 export const getManySubject = async () => {
   const { data } = await http.get<ISubject[]>('/subject')
@@ -67,4 +33,34 @@ export const editOneSubject = async (subject: ISubject) => {
 
 export const deleteOneSubject = async (id: string) => {
   await http.delete(`/subject/${id}`)
+}
+
+// ==================
+// Subject x Workload
+// ==================
+
+export interface ICompensated {
+  section: number
+  compensatedList: {
+    compensatedId: string
+    originalDate: string
+    originalTimeList: { start: string; end: string }[]
+    compensatedDate: string
+    compensatedTimeList: { start: string; end: string }[]
+    originalRoom: string
+    compensatedRoom: string
+  }[]
+}
+
+export const getManyCompensatedOfSubject = async (
+  subjectId: string,
+  query: any
+) => {
+  const { data } = await http.get<ICompensated[]>(
+    `/subject/${subjectId}/compensation-workload`,
+    {
+      params: query,
+    }
+  )
+  return data
 }
